# Faster `sqrtm()` by leveraging broadcasting