@pierreguilmin 2024-05-01
> I don't get why that works, but ok haha

Yes, I was expecting max a x2 speedup (bc we save one matrix multiplication), but the results on CPU are stupidly good. @gaspardbb any hints? This does not scale as a full matrix multiplication, what's the catch?

E.g. on CPU `x @ x` scales as
```
=== n=16
5.46 µs ± 617 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
=== n=128
169 µs ± 3.3 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
=== n=256
1.13 ms ± 107 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```