@pierreguilmin 2024-05-01
Ok got it, in the proposed benchmark with `_sqrtm_new` JAX can pre-compute the matrix multiplication away during the JIT and never compute it again. Here's a fairer comparison of the two versions, there's still a gain but minor 😋 The best gain is probably memory-wise (no extra allocation).

```python
@jax.jit
def _sqrtm_old(x):
    w, v = jnp.linalg.eigh(x)
    w = jnp.where(w < 0, 0, w)
    return v @ jnp.diag(jnp.sqrt(w)) @ v.mT.conj()

@jax.jit
def _sqrtm_new(x):
    w, v = jnp.linalg.eigh(x)
    w = jnp.where(w < 0, 0, w)
    return (v * jnp.sqrt(w)[None, :]) @ v.mT.conj()
```

**CPU (MacBook Air M1)**
```
=== n=16
29 µs ± 1.26 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
27.2 µs ± 1.4 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
=== n=128
11.2 ms ± 3.31 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)
8.31 ms ± 2.08 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)
=== n=256
44.2 ms ± 20.1 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
36.9 ms ± 18.5 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

**GPU (NVIDIA L40s)**
```
=== n=16
468 µs ± 8.68 µs per loop (mean ± std. dev. of 10 runs, 100 loops each)
456 µs ± 5.57 µs per loop (mean ± std. dev. of 10 runs, 100 loops each)
=== n=128
1.56 ms ± 23 µs per loop (mean ± std. dev. of 10 runs, 100 loops each)
1.54 ms ± 19.1 µs per loop (mean ± std. dev. of 10 runs, 100 loops each)
=== n=256
2.08 ms ± 24.9 µs per loop (mean ± std. dev. of 10 runs, 100 loops each)
2.08 ms ± 16.7 µs per loop (mean ± std. dev. of 10 runs, 100 loops each)
=== n=1024
9.62 ms ± 106 µs per loop (mean ± std. dev. of 10 runs, 100 loops each)
9.34 ms ± 36.4 µs per loop (mean ± std. dev. of 10 runs, 100 loops each)
```

cc @gautierronan and @abocquet