@gaspardbb 2024-05-02
Nice! Indeed there seems to be a small gain on CPU?
My motivation is simply that I'm worried I'm allocating a full diagonal matrix when writing `jnp.diag(w)`, which might not be optimized away. Writing it `w[None, :]` does the same and relies on broadcasting so I find it safer. 

Note that broadcasting rule will insert a new axis on the left, so `(v * jnp.sqrt(w)) @ v.mT.conj()` does the same; or use `v @ (jnp.sqrt(w)[:, None] * v.mT.conj())`.